Add types to loada and storea pointer parameters.
The drawback here is that you are more likely to have to cast, but IMHO
it is *well* worth it.  We gain more type safety since you can't
accidentally store to the wrong pointer type (i.e., wav_i32x4_t ->
float*).  We can also use conformant array parameters, and the API is a
bit more self-documenting.