Move control parameter of bitselect and blend to the beginning.
This doesn't match wasm_simd128.h, but I feel that this order is more
natural since it matches the ternary operator from standard C:
`control ? when_set : when_unset`.

The orders for different APIs:

 * wasm_simd128.h: when_set, when_unset, control
 * x86 (blendv): when_set, when_unset, control
 * NEON: control, when_set, when_unset
 * SVE: control, when_set, when_unset
 * AltiVec: when_set, when_unset, control

I don't know what the function is in MVE, but I suspect control may be
at the end since in MVE predicates come last.

So this *might* be a mistake; people used to blendv might be surprised.
Matching NEON is nice, though there aren't currently as many people
familiar with NEON as with SSE 4.1.  However, I think this order makes
more sense and this is a good opportunity to switch.